+++ webide2/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+execution-server-uri=http://localhost:8081/exec
